+++ pkg/azuremachinepool/validate_azuremachinepool_update.go
@@ -49,6 +49,11 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return microerror.Maskf(parsingFailedError, "unable to parse azureMachinePool CR: %v", err)
 	}
 
+	if !azureMPNewCR.GetDeletionTimestamp().IsZero() {
+		a.logger.LogCtx(ctx, "level", "debug", "message", "The object is being deleted so we don't validate it")
+		return nil
+	}
+
 	capi, err := generic.IsCAPIRelease(azureMPNewCR)
 	if err != nil {
 		return microerror.Mask(err)