+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Don't validate objects if they are being deleted.
+
 ### Added
 
 - Skipping all validation and defaulting for resources with CAPI release label